Redo parsing of meta data
The meta data collection is simple. Hunt for GAME=foo and note the
comment character(s) that appeared before it. This requires a comment
string before it, even if the language supports multiline comments. Take
that comment string and search for all KEY=VALUE strings, requiring KEY
to be in \w. Simple, a bit naive maybe, but good enough.