Add a passed column to result table and check if the execution passed
The part doing the execution has all of the data on whether execution of
the build or run passed so use it. The output and expected output are
stripped of leading and trailing whitespace and then compared. Only then
will a run be considered to have passed. For builds, the build process
needs to return an exit status of zero.