Use a UUID for the run id instead of an integer
If using an integer, then it would be much harder to merge separate
databases together. Switching to UUID version 4 means every run will
have a unique identifier. Full or partial runs can be done with a
non-existent database and then adding the rows from that database back
to a "main" database with all the history.