Use a namedtuple for the result
Join the result table with run to get the impl, game, and time it was
run. Stuff all of that into a namedtuple. This makes the code look a lot
better than before.

Also, only rank based on total CPU time taken. Don't do a mix of
results. It could be possible to cheat at the system by having one run
maximum memory usage to minimize CPU and another run that does the
opposite. The code before would pick the minimum of both while the new
code will take whatever the fastest run did.